+++ lib/build.nix
@@ -1,8 +1,9 @@
-{ build, system, version, lib, util, inputs, pkgs, files, ... }:
+{ build, system, version, lib, util, inputs, pkgs, patches, files, ... }:
 with inputs;
 let
   inherit (self) nixosModules;
   specialArgs = { inherit util inputs files; };
+  nixosSystem = import (patches + "/nixos/lib/eval-config.nix");
 
   # Shared Configuration Modules
   common =
@@ -26,7 +27,7 @@ in
       nixosModules.secrets
       nixosModules.shell
     ];
-  in lib.nixosSystem
+  in nixosSystem
   {
     inherit system specialArgs;
     modules =
@@ -63,7 +64,7 @@ in
   let
     # Install Media Build Module
     module = [ "${nixpkgs}/nixos/modules/installer/cd-dvd/iso-image.nix" ];
-  in lib.nixosSystem
+  in nixosSystem
   {
     inherit system specialArgs;
     modules =